Move ComponentVariants interface to public API
