+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/ComponentVariants.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.platform.base.internal;
+package dev.nokee.platform.base;
 
+/**
+ * The variants for a component.
+ *
+ * @since 0.5
+ */
 public interface ComponentVariants {
 }
