+++ subprojects/platform-jni/src/main/java/dev/nokee/platform/jni/internal/JavaNativeInterfaceComponentVariants.java
@@ -24,6 +24,7 @@ import dev.nokee.language.swift.SwiftSourceSet;
 import dev.nokee.model.internal.DomainObjectEventPublisher;
 import dev.nokee.model.internal.core.ModelSpecs;
 import dev.nokee.model.internal.registry.ModelLookup;
+import dev.nokee.platform.base.ComponentVariants;
 import dev.nokee.platform.base.internal.*;
 import dev.nokee.platform.base.internal.binaries.BinaryViewFactory;
 import dev.nokee.platform.base.internal.dependencies.ConfigurationBucketRegistryImpl;
